stop if git tags were updated, allow non-force autoconf, enhanced build speed, include autostart instruction in postinst and adjusted start hint to systemd

+++ autogen.sh
@@ -4,5 +4,10 @@ test -n "$srcdir" || srcdir=.
 
 mkdir -p "$srcdir/build"
 
-autoreconf --force --install --verbose "$srcdir"
+if [ "x$1" = "x--dontforce" ]; then
+  shift
+  autoreconf --install --verbose "$srcdir"
+else
+  autoreconf --force --install --verbose "$srcdir"
+fi
 test -n "$NOCONFIGURE" || "$srcdir/configure" "--prefix=/usr" "--sysconfdir=/etc" "--localstatedir=/var" "$@"
